+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 
  - Added "userFavoriteIds" variable in CategoryController::viewAction
  - Upgraded SDK to v1.14.0
+ - Added `\WizaplaceFrontBundle\Service\AuthenticationService::getInitiateResetPasswordForm`
+ - Added `\WizaplaceFrontBundle\Service\AuthenticationService::initiateResetPassword`
+ - Deprecated `\WizaplaceFrontBundle\Controller\AuthController::initiateResetPasswordAction`
 
 ### Bugfixes